frontend/agentverse-streamlit-app/requirements.txt
streamlit
neo4j
plotly
pyvis
networkx
python-dotenv
pandas